+++ openadapt/strategies/visual.py
@@ -189,7 +189,7 @@ def get_action_prompt_dict(action: models.ActionEvent) -> dict:
         }
     )
     if action.active_segment_description:
-        for key in ("mouse_x", "mouse_y", "mouse_dx", "mouse_dy"):
+        for key in ("mouse_x", "mouse_y", "mouse_dx", "mouse_dy", "reducer_names"):
             if key in action_dict:
                 del action_dict[key]
     if action.available_segment_descriptions:
@@ -207,13 +207,13 @@ class VisualReplayStrategy(
     def __init__(
         self,
         recording: models.Recording,
-        replay_instructions: str,
+        instructions: str,
     ) -> None:
         """Initialize the VisualReplayStrategy.
 
         Args:
             recording (models.Recording): The recording object.
-            replay_instructions (str): Natural language instructions for how recording
+            instructions (str): Natural language instructions for how recording
                 should be replayed.
         """
         super().__init__(recording)
@@ -221,7 +221,7 @@ class VisualReplayStrategy(
         add_active_segment_descriptions(recording.processed_action_events)
         self.modified_actions = apply_replay_instructions(
             recording.processed_action_events,
-            replay_instructions,
+            instructions,
         )
         global DEBUG
         DEBUG = DEBUG_REPLAY